pyproject.toml
[project]
name = "logistics-rpf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pandas>=3.0.1",
    "pip>=26.0.1",
    "python-dotenv>=1.2.2",
    "regex>=2026.2.28",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.3.0",
    "spacy>=3.8.11",
    "thefuzz>=0.22.1",
]
